// src/app/generated/blueprint.api/model/proficiencyScale.ts
/**
 * Blueprint API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProficiencyLevel } from './proficiencyLevel';


export interface ProficiencyScale { 
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    name?: string | null;
    description?: string | null;
    proficiencyLevels?: Array<ProficiencyLevel> | null;
}